+++ public/index.html
@@ -26,6 +26,7 @@
   <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
   <link href="https://fonts.googleapis.com/css2?family=Dancing+Script:wght@700&family=Inter&display=swap"
     rel="stylesheet">
+  <link rel="stylesheet" href="./style.css">
   <title>Teratany - SDG & Social Network</title>
 </head>
 
@@ -43,5 +44,10 @@
       To create a production bundle, use `npm run build` or `yarn build`.
     -->
 </body>
+<style>
+  .absolute.inset-0.w-full.h-full.pointer-events-auto.bg-black.bg-opacity-60.backdrop-blur-sm {
+    background-color: rgba(0, 0, 0, 0.521) !important;
+  }
+</style>
 
 </html>
